requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
Django==5.2.6
django-redis==6.0.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.9
PyYAML==6.0.2
redis==6.4.0
redis-lock==0.2.0
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
